Add ccdKey as a config option
This was previously hard-coded to 'ccd' but
obs_lsst uses 'detector' so this is now configurable.
Note that this take doesn't run from the command line
so there is no point in adding a config file to obs_lsst as
that magic is picked up by the argparser. Users will need to
know to set this in the config by hand.