Write a task to calculate, fit, and plot a Photon Transfer Curve
This task takes a list of pairs of flat fields in the command line
and calculates the Photon Transfer Curve from the difference of each
pair (var vs mean). The task fits different models to the PTC,
returning parameters such as gain and readout noise. The nonlinearity
residual is also calculated from a quadratic fit to the mean signal vs
exposure time curve.