Implement Gen3 version of calibration combine code.
The functions and tasks implemented here are taken directly from the
pipe_drivers constructCalibs.py code.  This version strips out the
ctrl_pool handling in favor of the Gen3 pipeline multiprocessing
handling.

These functions have not been fully tested with large numbers of
inputs, and may need to be reworked for memory usage.