Move to using the PTC task to calculate gains
Ensure gains are applied for all relevant code paths.
Update to use the newly re-factored PTC task (DM-22405).
Also, prior to this commit, the measured gains were not always applied.
Now the PTC is plotted for all paths if selected and available.
When gains are calculated using the PTC task, store them in the kernel.
Update the _applyGains() method to use the new format data, which stores
the used/unused visits.