Stop bad amps from causing total failure.
Add ability to mark amps as bad rather than just raising.
Also, record which visits are used as input, and add method to allow
easy access to which are actually used for a given amp, as the
mask is not trivially applied due to typing.

Add test for bad amp marking.