Add flat scaling tasks and pipeline definition.
Constructing flats for cameras with multiple detectors requires
scaling the individual exposures to a common flux level.
CpFlatMeasureTask measures statistics about the individual input
detectors and amplifiers, to allow the scale factors to be
constructed.

The output from all CpFlatMeasureTask runs are used in
CpFlatNormalizationTask to determine the scales to apply to the
inputs.  These scales can then be used by CalibCombineTask.

The full pipeline containing all these steps is in cpFlat.yaml.  A
second pipeline for cameras that have only a single detector, allowing
the intermediate scaling steps to be skipped, is in
cpFlatSingleChip.yaml.