Add initial code to certify a gen3 calibration.
Calibrations generated by CpCombineTask are not available for general
use.  A date range in which the calibration is valid must be manually
set and a calibration collection selected.  The new script provides
the command line interface to set the valid dates and final
collection.  The calibration_label dimension is populated by this
certification process, making the newly certified calibration usable.