Work around need for static camera in flat normalization and crosstalk.
These tasks are trying to do something ill-defined according to our
data model: get a camera with no temporal information.  This workaround
should work for now, but we need a better solution, and I'm not sure if
that's a change to the tasks or a change to the data model.