Standardize ptcDataset noise units to electrons
This will introduce a new PTC version (1.9) which
standardizes the PTC noise attribute units to
electron. Previously, FULLCOVARIANCE fit mistakenly
output in electron^2. Changing the PTC verison
allows for new checks in ISR tasks to use the
correct noise measurement.

This previously caused the variance plane to be
if PTC version < 1.9 and ptcFitType == FULLCOVARIANCE.